+++ src/cdk-experimental/testing/testbed/index.ts
@@ -6,8 +6,5 @@
  * found in the LICENSE file at https://angular.io/license
  */
 
-export * from './component-harness';
-export * from './harness-environment';
-export * from './protractor';
-export * from './test-element';
-export * from './testbed';
+export * from './testbed-harness-environment';
+export * from './unit-test-element';
