+++ src/cdk-experimental/testing/tests/protractor.e2e.spec.ts
@@ -1,5 +1,4 @@
 import {browser} from 'protractor';
-
 import {ProtractorHarnessEnvironment} from '../protractor';
 import {MainComponentHarness, WrongComponentHarness} from './harnesses/main-component-harness';
